Users cannot switch to fr_FR
When you try to switch to the French language in the web UI it's like the
change is not taken into account.

Most of Tuleap is internationalized thanks to ext-gettext. ext-gettext
will use the locale that Tuleap via a call to \setlocale(). This call to
\setlocale() will in turn call setlocale(3) [0][1]. setlocale(3) only
do something if the provided locale value is a valid locale
specification. In our case since only the image is built only with some
en_XX locales, switching to French does do anything and the UI stays in
English.

Issue has been identified thanks to request #17969: End to End tests
should use CentOS 7

[0] https://github.com/php/php-src/blob/php-7.3.24/ext/standard/string.c#L36-L43
[1] https://github.com/php/php-src/blob/php-7.3.24/ext/standard/string.c#L4854